backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pydantic
google-generativeai
python-dotenv